Upload to unstable as 5.0.2-6
I would have waited for more changes for this upload, but since the
first signs of the next upstream release are already there, I'd like
to get the new Hurd FTBFS fix tested so that I can submit it to
upstream _before_ the new upstream release.

Hence that rather small diff for this upload.

BTW, I mixed up names in the previous commit message: It was Bart
Schaefer of zsh upstream who suggested the fix, not Bart Martens of
Debian. I'm sorry, but I noticed this just now when preparing the
Debian changelog.